% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{circuit_add_gate}
\alias{circuit_add_gate}
\title{Add a logic gate to the circuit.}
\usage{
circuit_add_gate(circuit, gate)
}
\arguments{
\item{circuit}{Circuit which will receive the gate.}

\item{gate}{Logic gate to be added into the circuit.}
}
\value{
The new circuit with the logic gate added.
}
\description{
Add a logic gate to the circuit.
}
