% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{circuit_insert_gate}
\alias{circuit_insert_gate}
\title{Inserts a gate into the circuit schema without pre-compilation of the circuit.}
\usage{
circuit_insert_gate(circuit, gate)
}
\arguments{
\item{circuit}{Circuit which will receive the gate.}

\item{gate}{Gate to be inserted into the circuit.}
}
\value{
The circuit with the inserted gate.
}
\description{
Inserts a gate into the circuit schema without pre-compilation of the circuit.
}
