% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{circuit_link_gate_signals_not}
\alias{circuit_link_gate_signals_not}
\title{Link two signals from logic gates already added into the circuit.
This function uses the result from NOT(outputSingal) to link into
the asked inputSignal.}
\usage{
circuit_link_gate_signals_not(circuit, outputSignal, inputSignal)
}
\arguments{
\item{circuit}{Circuit which will receive the gate.}

\item{outputSignal}{The output signal to link.}

\item{inputSignal}{The input signal to link.}
}
\value{
The new circuit with the both signals linked.
}
\description{
Link two signals from logic gates already added into the circuit.
This function uses the result from NOT(outputSingal) to link into
the asked inputSignal.
}
