% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{compile_circuit}
\alias{compile_circuit}
\title{Compiles the circuit by combining all the species, reactions, initial concentrations and reaction constants.
This enables the circuit to be simulated in DNAr.}
\usage{
compile_circuit(circuit)
}
\arguments{
\item{circuit}{Circuit to be compiled.}
}
\value{
Compiled circuit.
}
\description{
Compiles the circuit by combining all the species, reactions, initial concentrations and reaction constants.
This enables the circuit to be simulated in DNAr.
}
