% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{gate_get_input_speciesnames}
\alias{gate_get_input_speciesnames}
\title{Return a list containing all the input species from a gate.}
\usage{
gate_get_input_speciesnames(gate, inputNumber)
}
\arguments{
\item{gate}{Gate used to search the species.}

\item{inputNumber}{Number of input}
}
\value{
a list containing all the input species from a gate.
}
\description{
Return a list containing all the input species from a gate.
}
