% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{make_circuit}
\alias{make_circuit}
\title{Create an empty circuit.}
\usage{
make_circuit(timing)
}
\arguments{
\item{timing}{A vector consisting of initial time, final time and timesteps to do the circuit simulation.}
}
\value{
An empty circuit.
}
\description{
Create an empty circuit.
}
