% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{make_flipflopd}
\alias{make_flipflopd}
\title{Create an FlipFlop-D with initial input values.}
\usage{
make_flipflopd(name, dValue, eValue)
}
\arguments{
\item{name}{The FlipFlop-D name}

\item{dValue}{The initial value of Data signal.}

\item{eValue}{The initial value of Enable signal}
}
\value{
An FlipFlop-D gate with its name, species, specific CRN reactions, ki and ci constants.
}
\description{
Create an FlipFlop-D with initial input values.
}
