% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{make_generic2to1_element}
\alias{make_generic2to1_element}
\title{Create a 2-inputs and 1-output generic element.}
\usage{
make_generic2to1_element(
  name,
  typename,
  in1str,
  input1value,
  in2str,
  input2value,
  outstr
)
}
\arguments{
\item{name}{The logic gate name}

\item{typename}{The type name of element (Latch, Flipflop, AND, OR, XOR, ...)}

\item{in1str}{The string identifier of input1 signal.}

\item{input1value}{The initial value of input1.}

\item{in2str}{The string identifier of input2 signal.}

\item{input2value}{The initial value of input2.}

\item{outstr}{The string identifier of output signal.}
}
\value{
A generic logic gate with its name, species, CRN reactions, ki and ci constants.
}
\description{
Create a 2-inputs and 1-output generic element.
}
