% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{make_generic2to1_gate}
\alias{make_generic2to1_gate}
\title{Create a 2-inputs and 1-output generic logic gate.}
\usage{
make_generic2to1_gate(name, typename, input1value, input2value)
}
\arguments{
\item{name}{The logic gate name}

\item{typename}{The type name of logic gate (AND, OR, XOR, ...)}

\item{input1value}{The initial value of input1.}

\item{input2value}{The initial value of input2.}
}
\value{
A generic logic gate with its name, species, CRN reactions, ki and ci constants.
}
\description{
Create a 2-inputs and 1-output generic logic gate.
}
