% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{make_latchd}
\alias{make_latchd}
\title{Create an Latch-D with initial input values.}
\usage{
make_latchd(name, dValue, eValue)
}
\arguments{
\item{name}{The Latch-D name}

\item{dValue}{The initial value of Data signal.}

\item{eValue}{The initial value of Enable signal}
}
\value{
An Latch-D gate with its name, species, specific CRN reactions, ki and ci constants.
}
\description{
Create an Latch-D with initial input values.
}
