% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{make_majority_gate}
\alias{make_majority_gate}
\title{Create an MAJORITY gate with initial input values.}
\usage{
make_majority_gate(name, input1value, input2value, input3value)
}
\arguments{
\item{name}{The logic gate name}

\item{input1value}{The initial value of input1.}

\item{input2value}{The initial value of input2.}

\item{input3value}{The initial value of input3.}
}
\value{
An AND gate with its name, species, specific CRN reactions, ki and ci constants.
}
\description{
Create an MAJORITY gate with initial input values.
}
