% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{make_signal}
\alias{make_signal}
\title{Create signal to use in the circuit.}
\usage{
make_signal(name)
}
\arguments{
\item{name}{The signal name.}
}
\value{
A signal with its name, species, CRN reactions and ki constants.
}
\description{
Create signal to use in the circuit.
}
