% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GATE_LIB.R
\name{make_signal_species}
\alias{make_signal_species}
\title{Create the singals' names of DNA species.}
\usage{
make_signal_species(name)
}
\arguments{
\item{name}{The signal name.}
}
\value{
List of signals' names.
}
\description{
Create the singals' names of DNA species.
}
